Focus type specific assertions a bit more, and move redundant assertions up to core dsl
