+++ src/main/java/uk/org/webcompere/modelassert/json/dsl/nodespecific/BooleanNodeDsl.java
@@ -5,8 +5,6 @@ import uk.org.webcompere.modelassert.json.condition.HasValue;
 import uk.org.webcompere.modelassert.json.condition.PredicateWrappedCondition;
 import uk.org.webcompere.modelassert.json.dsl.Satisfies;
 
-import static uk.org.webcompere.modelassert.json.condition.Not.not;
-
 /**
  * Assertions for Boolean nodes
  * @param <A> the overall type of the assertion
@@ -28,21 +26,5 @@ public interface BooleanNodeDsl<A> extends Satisfies<A> {
     default A isFalse() {
         return satisfies(new PredicateWrappedCondition("Boolean", JsonNode::isBoolean,
                 new HasValue<>(JsonNode::asBoolean, false)));
-    }
-
-    /**
-     * Assert that the node is boolean
-     * @return the assertion for fluent assertions, with this condition added
-     */
-    default A isBoolean() {
-        return satisfies(new PredicateWrappedCondition("Boolean", JsonNode::isBoolean));
-    }
-
-    /**
-     * Assert that the node is not boolean
-     * @return the assertion for fluent assertions, with this condition added
-     */
-    default A isNotBoolean() {
-        return satisfies(not(new PredicateWrappedCondition("Boolean", JsonNode::isBoolean)));
     }
 }
