+++ src/main/java/uk/org/webcompere/modelassert/json/dsl/nodespecific/BooleanNodes.java
@@ -1,5 +1,7 @@
 package uk.org.webcompere.modelassert.json.dsl.nodespecific;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import uk.org.webcompere.modelassert.json.condition.PredicateWrappedCondition;
 import uk.org.webcompere.modelassert.json.dsl.Satisfies;
 import uk.org.webcompere.modelassert.json.dsl.SubsetDsl;
 
@@ -9,6 +11,11 @@ import uk.org.webcompere.modelassert.json.dsl.SubsetDsl;
  */
 public class BooleanNodes<A> extends SubsetDsl<A> implements BooleanNodeDsl<A> {
     public BooleanNodes(Satisfies<A> satisfies) {
-        super(satisfies);
+        super(isBoolean(satisfies));
+    }
+
+    private static <A> Satisfies<A> isBoolean(Satisfies<A> requirements) {
+        requirements.satisfies(new PredicateWrappedCondition("Boolean", JsonNode::isBoolean));
+        return requirements;
     }
 }
