+++ src/test/java/uk/org/webcompere/modelassert/json/dsl/nodespecific/ArrayNodeDslTest.java
@@ -1,23 +1,14 @@
 package uk.org.webcompere.modelassert.json.dsl.nodespecific;
 
 import org.junit.jupiter.api.Test;
+import org.opentest4j.AssertionFailedError;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
+import static uk.org.webcompere.modelassert.json.JsonAssertions.assertJson;
 import static uk.org.webcompere.modelassert.json.TestAssertions.assertAllWays;
 import static uk.org.webcompere.modelassert.json.condition.ConditionList.conditions;
 
 class ArrayNodeDslTest {
-
-    @Test
-    void canDetectArray() {
-        assertAllWays("[]", "{}",
-            ArrayNodeDsl::isArray);
-    }
-
-    @Test
-    void canDetectNonArray() {
-        assertAllWays("{}", "[]",
-            ArrayNodeDsl::isNotArray);
-    }
 
     @Test
     void isArrayContainingStrings() {
@@ -103,5 +94,24 @@ class ArrayNodeDslTest {
                 .hasValue(3)
                 .hasValue(2)
                 .hasValue(1)));
+    }
+
+    @Test
+    void whenUsingArrayNodeThenNodeMustBeArray() {
+        assertThatThrownBy(() -> assertJson("{foo:{}}")
+            .at("/foo").array().isEmpty())
+            .isInstanceOf(AssertionFailedError.class);
+    }
+
+    @Test
+    void whenUsingArrayNodeWithArrayThenIsEmptyWorks() {
+        assertJson("{foo:[]}")
+            .at("/foo").array().isEmpty();
+    }
+
+    @Test
+    void whenUsingArrayNodeWithArrayThenIsNotEmptyWorks() {
+        assertJson("{foo:[\"bar\"]}")
+            .at("/foo").array().isNotEmpty();
     }
 }
