+++ src/test/java/uk/org/webcompere/modelassert/json/dsl/nodespecific/TextNodeDslTest.java
@@ -1,24 +1,15 @@
 package uk.org.webcompere.modelassert.json.dsl.nodespecific;
 
 import org.junit.jupiter.api.Test;
+import org.opentest4j.AssertionFailedError;
 import uk.org.webcompere.modelassert.json.dsl.JsonNodeAssertDsl;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
+import static uk.org.webcompere.modelassert.json.JsonAssertions.assertJson;
 import static uk.org.webcompere.modelassert.json.Patterns.GUID_PATTERN;
 import static uk.org.webcompere.modelassert.json.TestAssertions.assertAllWays;
 
 class TextNodeDslTest {
-
-    @Test
-    void isText() {
-        assertAllWays("\"Damian\"", "null",
-                TextNodeDsl::isText);
-    }
-
-    @Test
-    void isNotText() {
-        assertAllWays("null", "\"Damian\"",
-                TextNodeDsl::isNotText);
-    }
 
     @Test
     void isTextWithValue() {
@@ -90,5 +81,24 @@ class TextNodeDslTest {
         assertAllWays("\"625110f5-502f-4748-8f1d-bad2237aa0aa\"",
             "\"not a guid\"",
             assertion -> assertion.textMatches("Has dashes", text -> text.contains("-")));
+    }
+
+    @Test
+    void whenUsingTextNodeNodeMustBeText() {
+        assertThatThrownBy(() -> assertJson("{foo:[]}")
+            .at("/foo").text().isEmpty())
+            .isInstanceOf(AssertionFailedError.class);
+    }
+
+    @Test
+    void whenUsingTextNodeWithTextIsEmptyWorks() {
+        assertJson("{foo:\"\"}")
+            .at("/foo").text().isEmpty();
+    }
+
+    @Test
+    void whenUsingTextNodeWithTextIsNotEmptyWorks() {
+        assertJson("{foo:\"bar\"}")
+            .at("/foo").text().isNotEmpty();
     }
 }
